Release 3.1.1
- Fix Android build break: Utils.java passed an int to Order.Amount.setValue,
  which payment-sdk 5.x changed to Double — cast the amount so it compiles.
  Without this, every consumer's Android build failed to compile against 5.2.x.
- Bump native payment-sdk 5.2.1 -> 5.2.2 (removes sensitive cURL/token Logcat
  logging from the HTTP client).
- SDK_VERSION / package version -> 3.1.1.

Android toolchain note: payment-sdk 5.2.x requires AGP 8 / Gradle 8 / compileSdk 35
/ JDK 17, i.e. React Native 0.71+. RN 0.70.x apps cannot build it.

+++ android/src/main/java/ae/network/Utils.java
@@ -60,7 +60,9 @@ public class Utils {
         // Need to revisit this and add a builder to make it look better
         Order.Amount amount = new Order.Amount();
         amount.setCurrencyCode(orderMap.getMap("amount").getString("currencyCode"));
-        amount.setValue(orderMap.getMap("amount").getInt("value"));
+        // payment-sdk-core 5.x changed Order.Amount.setValue to take a Double
+        // (was int-compatible in 3.0.x); cast so the int amount autoboxes correctly.
+        amount.setValue((double) orderMap.getMap("amount").getInt("value"));
         order.setAmount(amount);
 
         // Set links
